Update SocketIOClient.swift
With this change the swift library will have the same behavior of the Java version. When you whant to create a socket client that have namespace on Java you just create a Socket-IO object.

So, if you inform something like "http://your_ip:your_port/path"

It will use your path as a namespace.